video/requirements.txt
Flask==0.12.2
google-cloud-videointelligence==1.0.0
google-cloud-storage==1.6.0
gunicorn==19.7.1